Fix error when the package-and-class regex doesn’t match
When an import such as the following is encountered, the regular expression
doesn’t match:

    import somePackage.*

This will cause ‘s-match’ to return nil, which ‘cl-subseq’ doesn’t handle. This
change makes the ‘groovy-imports-get-package-and-class’ return nil when there is
no match instead of raise an error.

Alternatively the regular expression could be changed to support that import
statement, but in that case we add ‘*’ to the import cache. I don’t think that’s
something we want to have happen.